shear trajectory corrector
While shearing the simulation with periodic boundary conditions, some molecules are broken and shifted due to the shear (which is applied in the zx-plane in the x-direction). Thus, those broken molecules need to be fixed.

This script separates the whole and broken molecules and puts the broken molecules on the lattice boxes that are 27 copies of the original box around it by taking the shift due to the shear into account. Then, it detects the whole version of the broken molecule by calculating inter-particle distances and finally picks the target molecule for each broken molecule if its COM is inside the original box. It saves the final trajectory for all atoms and also for the COM of molecules in .txt format.

One can use this script for both all-atom and coarse-grained simulations but it needs some modifications for CG simulations as the anion is a single bead molecule and is always "whole"; thus, some lines need to be commented out in that case.