MSD calculator
This script calculates the mean squared displacement by reading the unwrapped COM trajectory of cations and anions resulting from the "unwrap_trajectory.py" script.

It is designed to calculate the MSD for all-atom simulations but can be easily adjusted for CG simulations as well. Moreover, it calculates the MSD in yz-plane but it can easily be adjusted for other directions.